# [fix] fix dsv3 weight loader tqdm and simplify shared experts fusion